Make translate own page and adjust style
