+++ app/web/components/LanguagePickerSettings.tsx
@@ -1,20 +1,12 @@
-import { ExpandMore as ExpandMoreIcon } from "@mui/icons-material";
-import {
-  Accordion,
-  AccordionDetails,
-  AccordionSummary,
-  Link,
-  Typography,
-} from "@mui/material";
+import {  Link, Typography } from "@mui/material";
 import { useTranslation } from "i18n";
 import { LANGUAGE_MAP } from "i18n/constants";
 import { GLOBAL } from "i18n/namespaces";
 import { useRouter } from "next/router";
+import { translateJobURL, translateRoute } from "routes";
 
+import Button from "./Button";
 import LanguagePickerSelect from "./LanguagePickerSelect";
-import TranslationProgress from "./TranslationProgress";
-
-const VOLUNTEER_PAGE_LINK = "https://couchers.org/volunteer/translator";
 
 interface ChangeLanguageProps {
   className?: string;
@@ -39,7 +31,7 @@ export default function LanguagePickerSettings({
         </Typography>
         <Typography variant="body1" paragraph>
           <Link
-            href={VOLUNTEER_PAGE_LINK}
+            href={translateJobURL}
             target="_blank"
             rel="noreferrer noopener"
             underline="hover"
@@ -48,49 +40,14 @@ export default function LanguagePickerSettings({
           </Link>
         </Typography>
         <LanguagePickerSelect displayMode="rect" />
-
-        <Accordion
-          sx={{
-            mt: 2,
-            boxShadow: "none",
-            backgroundColor: "transparent",
-            "&:before": {
-              display: "none",
-            },
-            "& .MuiAccordionSummary-root": {
-              backgroundColor: "transparent",
-              padding: 0,
-            },
-            "& .MuiAccordionDetails-root": {
-              backgroundColor: "transparent",
-              padding: 0,
-            },
-          }}
-        >
-          <AccordionSummary
-            expandIcon={<ExpandMoreIcon />}
-            aria-controls="translation-progress-content"
-            id="translation-progress-header"
-            sx={{
-              "& .MuiAccordionSummary-content": {
-                display: "flex",
-                alignItems: "center",
-                gap: 1,
-              },
-              "& .MuiAccordionSummary-expandIconWrapper": {
-                order: -1, // Move icon to the left
-                marginRight: 0,
-              },
-            }}
+        <Typography paragraph sx={{ mt: 2 }}>
+          <Button
+            // variant="primary"
+            onClick={() => router.push(translateRoute)}
           >
-            <Typography sx={{ fontWeight: "bold" }}>
-              {t("global:language_preference.view_translation_progress")}
-            </Typography>
-          </AccordionSummary>
-          <AccordionDetails>
-            <TranslationProgress />
-          </AccordionDetails>
-        </Accordion>
+            {t("global:language_preference.translation_progress.view_progress")}
+          </Button>
+        </Typography>
       </>
     </div>
   );
