+++ app/web/features/translate/TranslationPage.tsx
@@ -1,0 +1,7 @@
+import TranslationProgress from "./TranslationProgress";
+
+const TranslationPage = () => {
+  return <TranslationProgress />;
+};
+
+export default TranslationPage;
